requirements.txt
aiogram
python-dotenv
yt-dlp
load_dotenv
pyacoustid
chromaprint
aiohttp
musicbrainzngs
spotipy